Made some `Database` and business logic improvements

+++ WishList_Medium/ViewModels/WishViewModel.swift
@@ -11,21 +11,25 @@ import SwiftData
 @Observable
 class WishViewModel {
     // MARK: - Properties
-    var modelContext: ModelContext
+    var modelContext: ModelContext? = nil
     var wishItems = [WishItem]()
     
-    init(modelContext: ModelContext) {
+    init(modelContext: ModelContext? = nil) {
         self.modelContext = modelContext
         fetchWishItemData()
     }
     
     // MARK: - Methods
     func addItem(_ item: WishItem) {
+        guard let modelContext else { return }
+        
         modelContext.insert(item)
         fetchWishItemData()
     }
     
     func removeItem(_ indexSet: IndexSet) {
+        guard let modelContext else { return }
+        
         for index in indexSet {
             modelContext.delete(wishItems[index])
         }
@@ -33,6 +37,8 @@ class WishViewModel {
     }
     
     func fetchWishItemData() {
+        guard let modelContext else { return }
+        
         do {
             let descriptor = FetchDescriptor<WishItem>(sortBy: [SortDescriptor(\.date)])
             wishItems = try modelContext.fetch(descriptor)
